Dev: Added licence, strict types and a separate autoloader to allow running this locally
If the project has to be deployed, and composer requires a higher PHP version than locally available, running this command line tool locally will fail
since we are including the composer autoload file which checks that the PHP version installed is the same one Composer dependencies were installed with.

Also added function imports.